% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trac_api.R
\name{api_url}
\alias{api_url}
\title{Modify URL preserving existing path component}
\usage{
api_url(base_url, path = NULL, ...)
}
\arguments{
\item{base_url}{base url}

\item{path}{path to be appended, see \code{\link{modify_url}}}

\item{...}{other arguments passed to \code{\link{modify_url}}}
}
\value{
An updated URL
}
\description{
Modify URL preserving existing path component
}
\details{
\code{api_url} works just like \code{\link{modify_url}}, but when supplied 
the \code{path} argument it is appended to the existing path of
\code{base_url} instead of replacing it.
}
\seealso{
Other Trac accessors: \code{\link{get_trac_milestones}},
  \code{\link{get_trac_tickets}}, \code{\link{trac_api}},
  \code{\link{trac_login}}
}

