% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_gh_milestones.R
\name{fetch_gh_milestones}
\alias{fetch_gh_milestones}
\alias{fetch_gh_milestones.data.frame}
\title{Fetch milestones from GH}
\usage{
fetch_gh_milestones(object, ...)

\method{fetch_gh_milestones}{data.frame}(object, owner, repo, ...,
  verbose = TRUE, sleep = 1)
}
\arguments{
\item{object, }{R object, see Details}

\item{...}{other arguments passed to/from other methods}

\item{owner, repo}{GitHub owner and repo name}

\item{verbose}{logical, whether to provide more output}

\item{sleep}{numeric, delay between requests, passed to \code{Sys.sleep}}
}
\description{
Fetch milestones from GH
}

