% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trac_milestones.R
\name{get_trac_milestones}
\alias{get_trac_milestones}
\title{Get milestone metadata from Trac}
\usage{
get_trac_milestones(mnames, ...)
}
\arguments{
\item{mnames}{character, milestone names}

\item{...}{other arguments, currently ignored}
}
\value{
Data frame of milestone data with columns \code{trac_milestone_*} where \code{*} are:
\item{name}{milestone name}
\item{description}{milestone description}
\item{due_date}{milestone due date, if any}
\item{status}{milestone status, either "open" or "closed"}
\item{url}{URL to the milestone in Trac}
}
\description{
Get milestone metadata from Trac
}
\details{
\code{get_trac_milestones} fetches appropriate HTML pages from Trac and 
extracts neccessary data. The process is somewhat fragile with respect to the
user preferences set in Trac. In particular, it assumes that the time&date 
display is set to UTC.
}
\seealso{
Other Trac accessors: \code{\link{api_url}},
  \code{\link{get_trac_tickets}}, \code{\link{trac_api}},
  \code{\link{trac_login}}
}

