% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trac_tickets.R
\name{get_trac_tickets}
\alias{get_trac_tickets}
\title{Query Trac tickets}
\usage{
get_trac_tickets(..., base_url = getOption("trac2gh.base_url"))
}
\arguments{
\item{...}{key=value pairs to filter tickets}

\item{base_url}{base URL to Trac}
}
\value{
Data frame of ticket data.
}
\description{
Query Trac tickets
}
\seealso{
Other Trac accessors: \code{\link{api_url}},
  \code{\link{get_trac_milestones}},
  \code{\link{trac_api}}, \code{\link{trac_login}}
}

