% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_milestone_formatters.R
\name{gh_milestone_title}
\alias{gh_milestone_description}
\alias{gh_milestone_due_on}
\alias{gh_milestone_formatters}
\alias{gh_milestone_state}
\alias{gh_milestone_title}
\title{GitHub milestone formatters}
\usage{
gh_milestone_title(title)

gh_milestone_description(description, url, title, ...)

gh_milestone_due_on(due_date)

gh_milestone_state(state)
}
\arguments{
\item{title}{milestone title}

\item{description}{character, milestone description}

\item{url}{character, milestone URL in Trac}

\item{...}{other arguments passed to/from other methods}

\item{due_date}{character, milestone due date}

\item{state}{character, milestone state}
}
\value{
Unless otherwise noted, all function return character vectors.
}
\description{
Functions formatting GitHub milestone data. They are vectorized so multiple
milestones can be processed.
}
\details{
\code{gh_milestone_description} creates milestone description for GH by
concatenating Trac milestone description with Trac milestone URL.
}

