% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_issue.R
\name{link_issue}
\alias{link_issue}
\title{Linking issues in commit messages}
\usage{
link_issue(ticket_id, db = issues)
}
\arguments{
\item{ticket_id}{character string of the form \code{#123}}

\item{db}{data frame mapping tickets to issues, see Details}
}
\value{
String with ticket reference replaced with GH reference (if found in
\code{db}).
}
\description{
Trac ticket IDs are often referenced in git commit messages as e.g. \code{#123}.
Assuming that tickets have been converted to issues and pushed to GitHub this function
can be used in a script with \code{git filter-branch} to rewrite commit messages
and replacing ticket references with issue references.
}
\details{
The data frame must contain columns \code{trac_ticket_id} and
\code{gh_issue_number} containing Trac ticket ID and corresponding GH issue
number.

If the \code{ticket_id} number is found in \code{db} it is replaced with a
string \code{#XXX} where \code{XXX} is the corresponding issue number.
Original Trac ticket ID is also displayed in square brackets. If
\code{ticket_id} is not found in \code{db}, the original reference is
replaced with trac ticket info in square brackets.
}

