% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gh_issues.R
\name{make_gh_issues}
\alias{make_gh_issues}
\alias{make_gh_issues.data.frame}
\alias{make_gh_issues.default}
\title{Create data for GitHub issues}
\usage{
make_gh_issues(object, ...)

\method{make_gh_issues}{default}(object, ...)

\method{make_gh_issues}{data.frame}(object, milestones, ...)
}
\arguments{
\item{object}{R object, currently only a method for data frames}

\item{...}{other arguments}

\item{milestones}{data frame as returned by \code{\link{make_gh_milestones}}}
}
\value{
\code{make_gh_issues} returns a data frame with columns \code{gh_issue_*} with \code{*} as:
\item{title}{issue title}
\item{body}{issue body}
\item{milestone}{milestone number}
\item{label}{list column with vectors of issue labels}
}
\description{
Create data for GitHub issues
}
\details{
\code{make_gh_issues} is a generic function that transforms supplied data
to structure and format necessary for creating issues on GitHub.

\code{make_gh_issues} method for data frames assumes that \code{object} contains
Trac ticket metadata in columns as returned by \code{\link{get_trac_tickets}}.
}
\seealso{
\code{\link{gh_issue_formatters}}

Other Github converters: \code{\link{make_gh_milestones}}
}

