% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gh_milestones.R
\name{make_gh_milestones}
\alias{make_gh_milestones}
\alias{make_gh_milestones.data.frame}
\alias{make_gh_milestones.default}
\title{Create data for GitHub milestones}
\usage{
make_gh_milestones(object, ...)

\method{make_gh_milestones}{default}(object, ...)

\method{make_gh_milestones}{data.frame}(object, ...)
}
\arguments{
\item{object}{R object, currently only a method for data frames}

\item{...}{other arguments}
}
\value{
\code{make_gh_milestones} returns a data frame with columns
\code{gh_milestone_*} with \code{*} as:
\item{title}{milestone title}
\item{description}{milestone description}
\item{due_on}{milestone due date}
\item{state}{milestone status, either "open" or "closed"}
}
\description{
Create data for GitHub milestones
}
\details{
\code{make_gh_milestones} is a generic function that transforms supplied data
to structure and format necessary for creating milestones on GitHub.

\code{make_gh_milestones} method for data frames assumes that \code{object}
contains Trac milestone metadata in columns as returned by
\code{\link{get_trac_milestones}}.
}
\seealso{
\code{\link{gh_milestone_formatters}}

Other Github converters: \code{\link{make_gh_issues}}
}

