% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_script.R
\name{make_script}
\alias{make_script}
\title{Create script for git filter-branch}
\usage{
make_script(issue_data, outfile = "", fun = link_issue,
  script_template = NULL, ...)
}
\arguments{
\item{issue_data}{data frame with columns  mapping Trac tickets to GitHub issues, see Details}

\item{outfile}{output file name}

\item{fun}{function performing actual rewriting}

\item{script_template}{path to script template}

\item{...}{other arguments passed to \code{\link{whisker.render}}}
}
\value{
Nothing
}
\description{
Create script for git filter-branch
}
\details{
This function uses \pkg{whisker} templating mechanism to generate a script
file that can be used with \code{git filter-branch} to rewrite commit
messages based on Trac ticket to GitHub issues mapping.
}
\seealso{
\code{\link{link_issue}}
}

