% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moinmoin_to_markdown.R
\name{moinmoin_to_markdown}
\alias{moinmoin_to_markdown}
\title{Rudimentary MoinMoin wiki to Markdown converter}
\usage{
moinmoin_to_markdown(x)
}
\arguments{
\item{x}{character vector of texts to convert}
}
\value{
Translated text.
}
\description{
Rudimentary MoinMoin wiki to Markdown converter
}
\details{
At this moment handles fenced code blocks only.
}

