% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_gh_issues.R
\name{push_gh_issues}
\alias{push_gh_issues}
\alias{push_gh_issues.data.frame}
\title{Push issues to GitHub}
\usage{
push_gh_issues(object, ...)

\method{push_gh_issues}{data.frame}(object, owner, repo, ..., verbose = TRUE,
  sleep = 1, retval = c("df", "resplist", "dryrun"))
}
\arguments{
\item{object, }{R object}

\item{...}{other arguments passed to/from other methods}

\item{owner, repo}{character, GitHub repo and its owner, passed to \code{\link{create.issue}}}

\item{verbose}{logical, provide more output}

\item{sleep}{numeric, delay between subsequent requests, passed to \code{\link{Sys.sleep}}}

\item{retval}{character, one of "df", "resplist", or "dryrun", what should
the function return. See Value section.}
}
\value{
Depending on \code{retval} the function will return:
\item{df}{Data frame with columns from \code{object} with additional column
\code{gh_issue_number} with IDs assigned to issues by GitHub.}
\item{resplist}{list of unparsed responses from \code{GET}.}
\item{dryrun}{no requests to GH will be sent, return just a list of data that
are to be pushed.}
}
\description{
Push issues to GitHub
}

