% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_gh_milestones.R
\name{push_gh_milestones}
\alias{push_gh_milestones}
\alias{push_gh_milestones.data.frame}
\title{Push milestones to GH}
\usage{
push_gh_milestones(object, ...)

\method{push_gh_milestones}{data.frame}(object, owner, repo, ...,
  verbose = TRUE, sleep = 1)
}
\arguments{
\item{object}{R object, see Details for methods}

\item{...}{other arguments passed to \code{\link{create.milestone}}, \code{ctx} in particular}

\item{owner, repo}{repository and its owner, passed to \code{\link{create.milestone}}}

\item{verbose}{logical, provide more output}

\item{sleep}{numeric, passed to \code{\link{Sys.sleep}}, pause between requests}
}
\value{
The method for data frames return \code{object} with the following columns added:
\item{push_response}{list with responses returned by \code{\link{create.milestone}}}
\item{push_ok}{logical, whether the request was succesfully processed}
\item{gh_milestone_number}{numerical, milestone number as assigned by GH}
}
\description{
Push milestones to GH
}
\details{
Method for data frames make requests to GH API. The milestones are created
with data from columns of \code{object}. It is assumed, and checked, if the
\code{object} has columns \code{gh_milestone_*} where \code{*} are: title,
description, state, due_on.
}

