% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trac_api.R
\name{trac_api}
\alias{trac_api}
\title{Trac API}
\usage{
trac_api(path = NULL, query = NULL,
  base_url = getOption("trac2gh.base_url"), ...)
}
\arguments{
\item{path, query}{passed to \code{\link{modify_url}}}

\item{base_url}{base URL, defaults to \code{trac2gh.base_url} option}

\item{...}{other arguments passed to \code{\link{GET}}}
}
\value{
Unparsed response (from \code{\link{GET}})
}
\description{
Accessing Statnet Trac non-interactively from R.
}
\seealso{
Other Trac accessors: \code{\link{api_url}},
  \code{\link{get_trac_milestones}},
  \code{\link{get_trac_tickets}}, \code{\link{trac_login}}
}

