% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trac_api.R
\name{trac_login}
\alias{trac_login}
\title{Get the Trac login cookie}
\usage{
trac_login(path, usr = getenv("TRAC_USER"), psw = getenv("TRAC_PSW"),
  auth_type = "basic", ...)
}
\arguments{
\item{path}{path part of URL, passed to \code{\link{modify_url}}}

\item{usr, psw, auth_type}{user name, password and authentication type, passed
to \code{\link{authenticate}}}

\item{...}{other arguments passed to \code{\link{trac_api}}}
}
\value{
The response, as returned by \code{\link{GET}}, invisibly.
}
\description{
Get the Trac login cookie
}
\details{
Get the login cookie invisibly. Package \CRANpkg{httr} preserves cookies across
requests to the same website, so they will be used automatically.
}
\examples{
\dontrun{
trac_login("login")
}
}
\seealso{
Other Trac accessors: \code{\link{api_url}},
  \code{\link{get_trac_milestones}},
  \code{\link{get_trac_tickets}}, \code{\link{trac_api}}
}

